Aula - Testando o findById

+++ dscatalog/backend/src/main/java/com/devsuperior/dscatalog/services/ProductService.java
@@ -68,9 +68,9 @@ public class ProductService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-//        if (!productRepository.existsById(id)) {
-//            throw new ResourceNotFoundException("Recurso não encontrado");
-//        }
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         try {
             productRepository.deleteById(id);
         }
